pose_animator-master/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
mediapipe>=0.8.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.5.0
tensorboard>=2.8.0
tqdm>=4.60.0
scikit-learn>=1.0.0
scipy>=1.7.0
